Added illnesses and symptoms.
Added Illness class to be inherited by all possible illnesses listed by Dr. Duke.
Added a Symptoms enum to list down all possible reactions to the illness the patient may be facing.
Added a Diagnosis class to be used as a way to convert a list of observed symptoms to a list of possible illnesses the patient may be facing.