Added the getPossibleIllnesses function in Diagnosis.java
Added the getPossibleIllnesses function that takes in a list of symptoms observed by the patient and returns an ArrayList of possible illnesses the patient may be suffering from. These illnesses must match at least 50% of the symptoms observed for it to be listed in possibleIllnesses ArrayList, the output of the function.
Added the IllnessMatch.java class which contains the Illness candidate as well as the percentage of matching symptoms.
Fixed a typo in the Conjunctivitis.java for listing the wrong symptom.