Implement ingredients database and default recipes with Java resources
We need to be able to easily add new recipes without having
to write Java code.

We also need to be able to maintain a set of valid ingredients
so that we can provide helpful information to the user about
expected units of measure in recipes as well as providing
well-defined matchings from user ingredients to recipes.

Without these two resources, adding new recipes is hard
and users may not be able to match with recipes that use
slightly different names for ingredients.

Let's,
* create JSON resources for both known ingredients and
default recipes
* write tests to ensure that these resources are correctly
formatted and sound (no unknown ingredients in recipes)