Added assertions and logging.
Added an assertion in Diagnosis.java to prevent a divide by zero in the event the programmer created an Illness child with no symptoms.
Added logging to signal when diagnosis was ran, and what is the comparison value to each of the possible illnesses.
Fixed a typo on authorship @Author tag.