Added more symptoms the users can input.
Added 13 more options for the user to input when entering their symptoms.
Added a new function addSymptom to add users' symptoms to the symptoms ArrayList.
Added error handling in Patient.java in the event no medication is available for the illness.
Removed boilerplate lines in case statement when user inputs their symptoms and moved them to addSymptom
Fixed Insomnia.java to display the correct symptom.