Parser Class : edit parseViewCommand method
All strings that match the matcher patterns will be removed from the
original user input string. As such, strings with unique identifiers
will be matched first. The view count input will be initialised as an
empty String and will be converted to MAX_DOUBLE if not specified. This
view count input will be matched last since it does not have a unique
identifier.