Modify VALID_PRICE_REGEX in ParserConstants Class
VALID_PRICE_REGEX now only matches with [1-9 digits] . [<= 2 dp].

Add more tests in ParserTest Class for View, Edit and Add commands

To ensure that all invalid amount arguments are rejected regardless of
the command type.

Modified MESSAGE_INVALID_AMOUNT in MessageConstants Class

To give the users a more accurate description of the cause of this
error.